Make a draft solution for gutenberg publish
