+++ WordPress/Classes/ViewRelated/Post/PostEditor.swift
@@ -4,14 +4,14 @@ import Gridicons
 
 /// Common interface to all editors
 ///
-@objc protocol PostEditor: class {
+protocol PostEditor: class {
     /// Initialize editor with a post.
     ///
     init(post: AbstractPost)
 
     /// The post being edited.
     ///
-    var post: AbstractPost { get }
+    var post: AbstractPost { get set }
 
     /// Closure to be executed when the editor gets closed.
     ///
@@ -20,9 +20,47 @@ import Gridicons
     /// Whether the editor should open directly to the media picker.
     ///
     var isOpenedDirectlyForPhotoPost: Bool { get set }
+
+    /// Boolean indicating whether the post should be removed whenever the changes are discarded, or not.
+    ///
+    var shouldRemovePostOnDismiss: Bool { get }
+    
+    /// Cancels all ongoing uploads
+    ///
+    ///TODO: We won't need this once media uploading is extracted to PostEditorUtil
+    func cancelUploadOfAllMedia(for post: AbstractPost)
+    
+    /// Whether the editor has failed media or not
+    ///
+    //TODO: We won't need this once media uploading is extracted to PostEditorUtil
+    var hasFailedMedia: Bool { get }
+
+    //TODO: We won't need this once media uploading is extracted to PostEditorUtil
+    var isUploadingMedia: Bool { get }
+
+    //TODO: We won't need this once media uploading is extracted to PostEditorUtil
+    //TODO: Otherwise the signature needs refactoring, it is too ambiguous for a protocol method
+    func removeFailedMedia()
+
+    /// Verification prompt helper
+    var verificationPromptHelper: VerificationPromptHelper? { get }
+
+    /// Post editor state context
+    var postEditorStateContext: PostEditorStateContext { get }
+
+    /// Update editor UI with given html
+    func setHTML(_ html: String)
+    
+    /// Return the current html in the editor
+    func getHTML() -> String
+
+    /// Title of the post
+    var postTitle: String { get set }
+    
+    var navigationBarManager: PostEditorNavigationBarManager { get }
 }
 
-protocol PostEditorNavigationBarManagerDelegate: class {
+protocol PostEditorNavigationBarManagerDelegate: class { 
     var publishButtonText: String { get }
     var isPublishButtonEnabled: Bool { get }
     var uploadingButtonSize: CGSize { get }
